refactor(week2-task1): convert file names to kebab case
